Improve error messages
The algorithm is slightly changed. Now, when we find an error, we also update the internal state. So for example, if we find an invalid `use`, we report an error, and then proceed as if use is properly placed.

This should in general reduce the amount of noise in the error report at the expense of hiding some potential errors.