Improve traversal
Traversal was going too deep into the ast, which caused some false positives. For example, if a module attribute was used inside a function, traversal would interpret it as a module declaration, and report an error.

This commit changes the traversal to stop going deeper if the part is recognized. In addition, def, defp, defmacro, defmacrop, and defguard are also treated as module parts, to make sure that traversal doesn't go into the function body.