Use dialyxir in all envs
Running dialyzer in prod env can sometimes be beneficial, because there might exist some code which is compiled only in prod env (e.g. if `if Mix.env() == :prod` is used to conditionally define an alternative implementation).

Note that dialyzer is marked with `runtime: false`, so it won't be bundled into the OTP release.