Fix crash in changeset_error
Localdrive team reported a crash, which was caused by

```
errors: [
    email: {"has already been taken",
     [validation: :unsafe_unique, fields: [:email]]}
  ]
```

in the changeset.

It appears that `unsafe_validate_unique` adds some rich metadata (in this case `fields`) which will casue `to_string` to raise an error.

As an easy fix, I'm ignoring any fields which are not referenced in the string.