Move skafolder into a separate project
Skafolder will be provided as a mix archive which can be globally installed. To make this work correctly, we need to extract skafolder code into a separate project. This is the similar approach taken by phoenix, which includes its phx.new generator in a separate subproject under the phoenix repo.